+++ test/parallel/test-whatwg-url-custom-global.js
@@ -1,5 +1,7 @@
 'use strict';
 
+// Tests below are not from WPT.
+
 require('../common');
 const assert = require('assert');
 const { URL, URLSearchParams } = require('url');
