+++ test/parallel/test-whatwg-url-custom-searchparams-values.js
@@ -1,10 +1,11 @@
 'use strict';
 
+// Tests below are not from WPT.
+
 const common = require('../common');
 const assert = require('assert');
 const URLSearchParams = require('url').URLSearchParams;
 
-// Tests below are not from WPT.
 const params = new URLSearchParams('a=b&c=d');
 const values = params.values();
