+++ test/parallel/test-whatwg-url-custom-searchparams.js
@@ -1,11 +1,12 @@
 'use strict';
 
+// Tests below are not from WPT.
+
 const common = require('../common');
 const assert = require('assert');
 const { URL, URLSearchParams } = require('url');
 const fixtures = require('../common/fixtures');
 
-// Tests below are not from WPT.
 const serialized = 'a=a&a=1&a=true&a=undefined&a=null&a=%EF%BF%BD' +
                    '&a=%EF%BF%BD&a=%F0%9F%98%80&a=%EF%BF%BD%EF%BF%BD' +
                    '&a=%5Bobject+Object%5D';
