+++ test/parallel/test-whatwg-url-custom-tostringtag.js
@@ -1,10 +1,11 @@
 'use strict';
 
+// Tests below are not from WPT.
+
 require('../common');
 const assert = require('assert');
 const URL = require('url').URL;
 
-// Tests below are not from WPT.
 const toString = Object.prototype.toString;
 
 const url = new URL('http://example.org');
