+++ test/parallel/test-whatwg-url-searchparams-append.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const common = require('../common');
-const assert = require('assert');
+require('../common');
 const URLSearchParams = require('url').URLSearchParams;
 const { test, assert_equals, assert_true } = require('../common/wpt');
 
@@ -48,34 +47,3 @@ test(function() {
     assert_equals(params.get('first'), '1', 'Search params object has name "first" with value "1"');
 }, 'Append multiple');
 /* eslint-enable */
-
-// Tests below are not from WPT.
-{
-  const params = new URLSearchParams();
-  common.expectsError(() => {
-    params.append.call(undefined);
-  }, {
-    code: 'ERR_INVALID_THIS',
-    type: TypeError,
-    message: 'Value of "this" must be of type URLSearchParams'
-  });
-  common.expectsError(() => {
-    params.append('a');
-  }, {
-    code: 'ERR_MISSING_ARGS',
-    type: TypeError,
-    message: 'The "name" and "value" arguments must be specified'
-  });
-
-  const obj = {
-    toString() { throw new Error('toString'); },
-    valueOf() { throw new Error('valueOf'); }
-  };
-  const sym = Symbol();
-  assert.throws(() => params.set(obj, 'b'), /^Error: toString$/);
-  assert.throws(() => params.set('a', obj), /^Error: toString$/);
-  assert.throws(() => params.set(sym, 'b'),
-                /^TypeError: Cannot convert a Symbol value to a string$/);
-  assert.throws(() => params.set('a', sym),
-                /^TypeError: Cannot convert a Symbol value to a string$/);
-}
