+++ test/parallel/test-whatwg-url-searchparams-get.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const common = require('../common');
-const assert = require('assert');
+require('../common');
 const URLSearchParams = require('url').URLSearchParams;
 const { test, assert_equals, assert_true } = require('../common/wpt');
 
@@ -33,31 +32,3 @@ test(function() {
     assert_equals(params.get('fourth'), null, 'Search params object has no "fourth" name and value.');
 }, 'More get() basics');
 /* eslint-enable */
-
-// Tests below are not from WPT.
-{
-  const params = new URLSearchParams();
-  common.expectsError(() => {
-    params.get.call(undefined);
-  }, {
-    code: 'ERR_INVALID_THIS',
-    type: TypeError,
-    message: 'Value of "this" must be of type URLSearchParams'
-  });
-  common.expectsError(() => {
-    params.get();
-  }, {
-    code: 'ERR_MISSING_ARGS',
-    type: TypeError,
-    message: 'The "name" argument must be specified'
-  });
-
-  const obj = {
-    toString() { throw new Error('toString'); },
-    valueOf() { throw new Error('valueOf'); }
-  };
-  const sym = Symbol();
-  assert.throws(() => params.get(obj), /^Error: toString$/);
-  assert.throws(() => params.get(sym),
-                /^TypeError: Cannot convert a Symbol value to a string$/);
-}
