+++ test/parallel/test-whatwg-url-searchparams-getall.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const common = require('../common');
-const assert = require('assert');
+require('../common');
 const URLSearchParams = require('url').URLSearchParams;
 const { test, assert_equals, assert_true, assert_array_equals } =
   require('../common/wpt');
@@ -38,31 +37,3 @@ test(function() {
     assert_array_equals(matches, ['one'], 'Search params object has expected name "a" values');
 }, 'getAll() multiples');
 /* eslint-enable */
-
-// Tests below are not from WPT.
-{
-  const params = new URLSearchParams();
-  common.expectsError(() => {
-    params.getAll.call(undefined);
-  }, {
-    code: 'ERR_INVALID_THIS',
-    type: TypeError,
-    message: 'Value of "this" must be of type URLSearchParams'
-  });
-  common.expectsError(() => {
-    params.getAll();
-  }, {
-    code: 'ERR_MISSING_ARGS',
-    type: TypeError,
-    message: 'The "name" argument must be specified'
-  });
-
-  const obj = {
-    toString() { throw new Error('toString'); },
-    valueOf() { throw new Error('valueOf'); }
-  };
-  const sym = Symbol();
-  assert.throws(() => params.getAll(obj), /^Error: toString$/);
-  assert.throws(() => params.getAll(sym),
-                /^TypeError: Cannot convert a Symbol value to a string$/);
-}
