+++ test/parallel/test-whatwg-url-searchparams-has.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const common = require('../common');
-const assert = require('assert');
+require('../common');
 const URLSearchParams = require('url').URLSearchParams;
 const { test, assert_false, assert_true } = require('../common/wpt');
 
@@ -36,31 +35,3 @@ test(function() {
     assert_false(params.has('first'), 'Search params object has no name "first"');
 }, 'has() following delete()');
 /* eslint-enable */
-
-// Tests below are not from WPT.
-{
-  const params = new URLSearchParams();
-  common.expectsError(() => {
-    params.has.call(undefined);
-  }, {
-    code: 'ERR_INVALID_THIS',
-    type: TypeError,
-    message: 'Value of "this" must be of type URLSearchParams'
-  });
-  common.expectsError(() => {
-    params.has();
-  }, {
-    code: 'ERR_MISSING_ARGS',
-    type: TypeError,
-    message: 'The "name" argument must be specified'
-  });
-
-  const obj = {
-    toString() { throw new Error('toString'); },
-    valueOf() { throw new Error('valueOf'); }
-  };
-  const sym = Symbol();
-  assert.throws(() => params.has(obj), /^Error: toString$/);
-  assert.throws(() => params.has(sym),
-                /^TypeError: Cannot convert a Symbol value to a string$/);
-}
