+++ test/parallel/test-whatwg-url-searchparams-set.js
@@ -1,7 +1,6 @@
 'use strict';
 
-const common = require('../common');
-const assert = require('assert');
+require('../common');
 const URLSearchParams = require('url').URLSearchParams;
 const { test, assert_equals, assert_true } = require('../common/wpt');
 
@@ -34,34 +33,3 @@ test(function() {
     assert_equals(params.get('a'), '4', 'Search params object has name "a" with value "4"');
 }, 'URLSearchParams.set');
 /* eslint-enable */
-
-// Tests below are not from WPT.
-{
-  const params = new URLSearchParams();
-  common.expectsError(() => {
-    params.set.call(undefined);
-  }, {
-    code: 'ERR_INVALID_THIS',
-    type: TypeError,
-    message: 'Value of "this" must be of type URLSearchParams'
-  });
-  common.expectsError(() => {
-    params.set('a');
-  }, {
-    code: 'ERR_MISSING_ARGS',
-    type: TypeError,
-    message: 'The "name" and "value" arguments must be specified'
-  });
-
-  const obj = {
-    toString() { throw new Error('toString'); },
-    valueOf() { throw new Error('valueOf'); }
-  };
-  const sym = Symbol();
-  assert.throws(() => params.append(obj, 'b'), /^Error: toString$/);
-  assert.throws(() => params.append('a', obj), /^Error: toString$/);
-  assert.throws(() => params.append(sym, 'b'),
-                /^TypeError: Cannot convert a Symbol value to a string$/);
-  assert.throws(() => params.append('a', sym),
-                /^TypeError: Cannot convert a Symbol value to a string$/);
-}
