+++ test/parallel/test-whatwg-url-setters.js
@@ -6,13 +6,9 @@ if (!common.hasIntl) {
   common.skip('missing Intl');
 }
 
-const assert = require('assert');
 const URL = require('url').URL;
 const { test, assert_equals } = require('../common/wpt');
 const fixtures = require('../common/fixtures');
-
-const additionalTestCases =
-  require(fixtures.path('url-setter-tests-additional.js'));
 
 const request = {
   response: require(fixtures.path('url-setter-tests'))
@@ -80,48 +76,3 @@ function runURLSettersTests(all_test_cases) {
 
 startURLSettersTests()
 /* eslint-enable */
-
-// Tests below are not from WPT.
-
-{
-  for (const attributeToBeSet in additionalTestCases) {
-    if (attributeToBeSet === 'comment') {
-      continue;
-    }
-    const testCases = additionalTestCases[attributeToBeSet];
-    for (const testCase of testCases) {
-      let name = `Setting <${testCase.href}>.${attributeToBeSet}` +
-                 ` = "${testCase.new_value}"`;
-      if ('comment' in testCase) {
-        name += ` ${testCase.comment}`;
-      }
-      test(function() {
-        const url = new URL(testCase.href);
-        url[attributeToBeSet] = testCase.new_value;
-        for (const attribute in testCase.expected) {
-          assert_equals(url[attribute], testCase.expected[attribute]);
-        }
-      }, `URL: ${name}`);
-    }
-  }
-}
-
-{
-  const url = new URL('http://example.com/');
-  const obj = {
-    toString() { throw new Error('toString'); },
-    valueOf() { throw new Error('valueOf'); }
-  };
-  const sym = Symbol();
-  const props = Object.getOwnPropertyDescriptors(Object.getPrototypeOf(url));
-  for (const [name, { set }] of Object.entries(props)) {
-    if (set) {
-      assert.throws(() => url[name] = obj,
-                    /^Error: toString$/,
-                    `url.${name} = { toString() { throw ... } }`);
-      assert.throws(() => url[name] = sym,
-                    /^TypeError: Cannot convert a Symbol value to a string$/,
-                    `url.${name} = ${String(sym)}`);
-    }
-  }
-}
